Refactor code
Refactored code into three files. Class PermrowcolSynthesis in
file perm_row_col_synthesis.py gets CouplingMap upon
initialization and calls class PermRowCol method perm_row_col
upon method run. Class PermRowCol in file permrowcol.py contains
methods perm_row_col that contains functionality to run
permrowcol algorithm. File graph_utils.py contains helper
functions to handle graphs.
